Rerun connector until a full traversal is completed.
When a long running operation from a previous run has not completed, the
connector will exit without completing a traversal. In such cases,
updateItemMetadataTest() will fail because the items "removed" from the
fake repository will not be deleted by the connector and will continue
showing in the Indexing API.

The solution is to use the StatsManager to keep running the connector
until it successfully completes a full traversal.

Bug: 123352680

Change-Id: I26a7e26dd26c00ef7a184f76cdde9c5bdf86243d