Ban Truth from the IT common library.
Truth.assertThat() causes sporadic NoSuchMethodErrors due to a bug in
Truth itself, so use instead JUnit assertions.

I can't remove the dependency on Truth completely because the identity
integration test relies on some Truth methods that have no equivalent on
JUnit 4, but limiting the scope of Truth to "test" helps to prevent it
from being used in the common library.

Bug: 123863881
Change-Id: If4279200a8769d1eaa248f49e963a805162a78f8