Helper class to verify serving.
To search as a user, it's necessary to provide the corresponding client
secrets and stored credentials. The test setup script will download
these from Cloud Storage and provide them to the test via properties (in
the same way the service account credentials are passed in).

See go/connectors-serving-test for instructions on how to generate the
credentials and set up the environment.

Change-Id: I6588425cf5a33ae3634dcbff12f0d63d30e12d0e