Move assertions out of the stubbings in IndexingServiceTest.
Other changes:

The following methods in TestingHttpTransport were called by
IndexingServiceTest, but not needed:

    addDeleteItemReqResp
    addPushItemReqResp
    addUnreserveItemsReqResp
    addUpdateItemReqResp

IndexingServiceImpl makes the corresponding API calls through the
BatchingIndexingService, which is mocked, so these calls never make it
to the transport.

Rename "updateItem" to "indexItem" in method names for consistency.
Rename some other test methods to follow the new conventions.

Change-Id: I1326d28d4b0576a414e419122a52143980a37791