Classes representing the items spec.
The items spec is a representation of the expected state of indexed
items.

Jackson's ObjectMapper is used to convert between POJO and JSON and viceversa
(see Spec.toJson() and Spec.fromJson()).

The CL is on the bigger side, but it's mostly boilerplate and I think it
makes sense to see it and submit it as a whole.

See design at go/connectors-verifier.

Change-Id: I51ebb9cf92795bae330a9e3f2d8b8b05c8bfd232