Support external groups (#34)
Add an identity connector to create external groups. Group data is
specified in a JSON file. Groups should be added to an identity source
dedicated to this purpose. This connector is intended to make it
easier to create groups, such as "Everyone" in the Windows
environment, that may appear in ACLs but are not mapped by other
identity connectors.

Add support in Acl.getGroupPrincipal for mapping group names to the
identity source used with the external groups connector. When a
connector is configured with a property pointing to the file used in
the external groups connector, and the identity source id where the
external groups are created, the getGroupPrincipal method will use the
separate identity source id for any groups found in the external
groups file.

Bug: 124328777