Add support to set context attributes while indexing an item into cloudsearch. (#45)
Add support to set context attributes while indexing an item in cloudsearch. These context attributes will be used to boost results when a relevant request context is passed in the search request.
Also implement a basic integration test to index 2 items with context attributes and check that they are searchable.